feat: add support for Fedora based Cayo
Recent testing demonstrated that we certainly have users who need support
for older x86-64 microarchtectures (CentOS 10 only supports v3 and newer).

We should consider moving Cayo back to Fedora as we don't face this limit
there.

ONE of the goals of Cayo was greater stability, but also many process
and design improvments. Using Fedora we would retain all our local
improvements, but return to some of the "package churn" which has been
discussed in other forums.

To address part of the "stability" goal, I'd suggest we investigate
shipping a kernel from a "longterm" branch rather than "stable" as
Fedora normally does. However, this PR does not address the kernel
question.

+++ build_files/01-common.sh
@@ -17,35 +17,40 @@ rm -rf /opt /usr/local
 ln -sf var/opt /opt
 ln -sf ../var/usrlocal /usr/local
 
-# /*
-# remove subscription manager
-# */
-dnf -y remove \
-  libdnf-plugin-subscription-manager \
-  python3-subscription-manager-rhsm \
-  subscription-manager \
-  subscription-manager-rhsm-certificates
+DIST=$(rpm -E %dist)
+if [[ "${DIST}" == ".el"* ]]; then
+  # /*
+  # remove subscription manager
+  # */
+  dnf -y remove \
+    libdnf-plugin-subscription-manager \
+    python3-subscription-manager-rhsm \
+    subscription-manager \
+    subscription-manager-rhsm-certificates
+
+  # /*
+  # enable CRB, EPEL and other repos
+  # */
+  dnf config-manager --set-enabled crb
+  dnf -y install epel-release
+  dnf -y upgrade epel-release
+else
+  dnf -y install dnf5-plugins
+fi
 
 # /*
 # remove any wifi support from base
 # */
 dnf -y remove \
-    atheros-firmware \
-    brcmfmac-firmware \
-    iwlegacy-firmware \
-    iwlwifi-dvm-firmware \
-    iwlwifi-mvm-firmware \
-    mt7xxx-firmware \
-    nxpwireless-firmware \
-    realtek-firmware \
-    tiwilink-firmware
-
-# /*
-# enable CRB, EPEL and other repos
-# */
-dnf config-manager --set-enabled crb
-dnf -y install epel-release
-dnf -y upgrade epel-release
+  atheros-firmware \
+  brcmfmac-firmware \
+  iwlegacy-firmware \
+  iwlwifi-dvm-firmware \
+  iwlwifi-mvm-firmware \
+  mt7xxx-firmware \
+  nxpwireless-firmware \
+  realtek-firmware \
+  tiwilink-firmware
 
 # /*
 # packages which are more or less what we'd find in CoreOS
